modbus_module/requirements.txt
python-dotenv==1.0.0
pika==1.3.2
requests==2.31.0
mysqlclient==2.1.1
mysql-connector-python
pymodbus==3.4.1